python-sd-service/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
flask>=2.3.0
flask-cors>=4.0.0
pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
safetensors>=0.3.0
# Note: xformers removed for GTX 1060 compatibility
